tests: settings: Reducing duplicated code in FS tests
Common code has been moved out of FS specific source into common
source files.
